Edited User Created Items
When a user creates their own item, it goes into the B4A "UserItems" database.  This is kept separate from the AllItems database.
When logging into the DM Screen, it loads the items based on the username.  These random items can now be used when creating a shop
or when creating a random shop.